requirements.txt
alabaster==0.7.12
asgiref==3.3.4
attrs==20.3.0
Babel==2.9.0
Cerberus==1.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-appconf==1.0.4
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-jsonfield==1.4.1
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-registration==3.1.2
django-templated-mail==1.1.1
django-user-accounts==3.0.4
django-webpack-loader==1.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docutils==0.16
idna==2.10
imagesize==1.2.0
itypes==1.2.0
Jinja2==2.11.3
jsonfield==3.1.0
Markdown==3.3.4
markdown-it-py==0.6.2
MarkupSafe==1.1.1
mdit-py-plugins==0.2.6
mysql==0.0.2
mysqlclient==2.0.3
myst-parser==0.13.6
oauthlib==3.1.1
packaging==20.9
Pillow==8.3.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.6.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
Sphinx==3.5.4
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
swapper==1.1.2.post1
uritemplate==3.0.1
urllib3==1.26.4
